clear pkg/proxy/port.go port_test.go file
